Remove cookie Config from the Session struct
* A Session should avoid having http.Cookie specific settings
such as the cookie Config. Use the CookieStore's cookie Config
when setting a new http.Cookie on a http.ResponseWriter
* Stop exposing the cookie config of each session to callers,
which would allow improper Config mutations after a cookie
has been issued (confusing)
* Reorganize cookie Config fields a bit
